backend/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
mangum==0.17.0

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# 向量数据库
chromadb==0.4.22

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP 客户端
httpx==0.26.0

# 腾讯云 COS SDK
cos-python-sdk-v5==1.9.30

# PPT 解析和生成
python-pptx>=0.6.23
lxml>=4.9.0
Pillow>=9.0.0

# 向量化
sentence-transformers==2.3.1

# 任务队列
celery==5.3.6

# 配置
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 工具
python-dotenv==1.0.0

# 测试
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0
